+++ R/utils.R
@@ -716,6 +716,24 @@ is_label_available <- function(x) {
   length(x) > 0 && !is.null(x$label)
 }
 
+#' @importFrom tidyr unnest_wider
+#' @importFrom purrr map
+#' @importFrom rlang set_names
+#' @noRd
+add_attachments <- function(x, form) {
+  media_type <- c("image", "video", "audio", "file")
+  cond <- form$type %in% media_type
+  nm <- unique(form$name[cond])
+  nm <- intersect(names(x), nm)
+  if (any(cond) && length(nm) > 0) {
+    vnm <- paste0(rev(nm), "_url")
+    x[["_attachments"]] <- map(x[["_attachments"]],
+                               ~ set_names(.x[["download_url"]], vnm))
+    x <- unnest_wider(x, "_attachments")
+  }
+  x
+}
+
 #' @noRd
 #' @importFrom dplyr select
 #' @importFrom labelled labelled
@@ -765,6 +783,7 @@ postprocess_data_ <- function(x, form, lang, select_multiple_label = FALSE, cn) 
   x <- extract_geotrace_(x, form)
   x <- extract_geoshape_(x, form)
   x <- type_convert_(x, form)
+  x <- add_attachments(x, form)
   x <- remove_list_cols(x)
   x <- add_missing_cols_(x, cn)
   x <- val_labels_from_form_(x = x, form = form, lang = lang)
@@ -864,6 +883,7 @@ set_names_from_varlabel_dm_ <- function(x) {
 #' @importFrom stats na.omit
 #' @noRd
 kobo_form_names_ <- function(form) {
+  media_type <- c("image", "video", "audio", "file")
   if ("choices" %in% names(form)) {
     res <- mutate(form, "value_name" =
                           case_when(type %in% "geopoint" ~ list(c("",
@@ -874,6 +894,7 @@ kobo_form_names_ <- function(form) {
                                                                   "_wkt")),
                                     type %in% "geotrace" ~ list(c("", "_wkt")),
                                     type %in% "geoshape" ~ list(c("", "_wkt")),
+                                    type %in% media_type ~ list(c("", "_url")),
                                     type %in% "select_multiple" ~ lapply(.data$choices,
                                                                          \(ch) c("", paste0("_", unique(ch$value_name)))),
                                     .default = list(""))) |>
@@ -889,6 +910,7 @@ kobo_form_names_ <- function(form) {
                                                                   "_wkt")),
                                     type %in% "geotrace" ~ list(c("", "_wkt")),
                                     type %in% "geoshape" ~ list(c("", "_wkt")),
+                                    type %in% media_type ~ list(c("", "_url")),
                                     .default = list(""))) |>
       unnest("value_name") |>
       distinct(.data$name, .data$value_name)